created fund-faucet command
