+++ cmd/root.go
@@ -1,17 +1,16 @@
 package cmd
 
 import (
-	"github.com/dymensionxyz/roller/cmd/fund_faucet"
 	"github.com/dymensionxyz/roller/cmd/migrate"
 	"github.com/dymensionxyz/roller/cmd/run"
 	"github.com/dymensionxyz/roller/cmd/services"
+	"github.com/dymensionxyz/roller/cmd/tx"
 	"github.com/dymensionxyz/roller/cmd/utils"
 	"os"
 
 	"github.com/dymensionxyz/roller/cmd/config"
 	da_light_client "github.com/dymensionxyz/roller/cmd/da-light-client"
 	"github.com/dymensionxyz/roller/cmd/keys"
-	"github.com/dymensionxyz/roller/cmd/register"
 	"github.com/dymensionxyz/roller/cmd/relayer"
 	"github.com/dymensionxyz/roller/cmd/sequencer"
 	"github.com/dymensionxyz/roller/cmd/version"
@@ -36,7 +35,6 @@ func Execute() {
 func init() {
 	rootCmd.AddCommand(config.Cmd())
 	rootCmd.AddCommand(version.Cmd())
-	rootCmd.AddCommand(register.Cmd())
 	rootCmd.AddCommand(da_light_client.DALightClientCmd())
 	rootCmd.AddCommand(sequencer.SequencerCmd())
 	rootCmd.AddCommand(relayer.Cmd())
@@ -44,6 +42,6 @@ func init() {
 	rootCmd.AddCommand(run.Cmd())
 	rootCmd.AddCommand(services.Cmd())
 	rootCmd.AddCommand(migrate.Cmd())
-	rootCmd.AddCommand(fund_faucet.Cmd())
+	rootCmd.AddCommand(tx.Cmd())
 	utils.AddGlobalFlags(rootCmd)
 }
