+++ cmd/tx/fund_faucet/fund_faucet.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"fmt"
 	"github.com/dymensionxyz/roller/cmd/consts"
+	"github.com/dymensionxyz/roller/cmd/tx/tx_utils"
 	"github.com/dymensionxyz/roller/cmd/utils"
 	"github.com/dymensionxyz/roller/config"
 	"github.com/dymensionxyz/roller/relayer"
@@ -24,13 +25,15 @@ func Cmd() *cobra.Command {
 			srcChannel, _, err := rly.LoadChannels()
 			if err != nil || srcChannel == "" {
 				utils.PrettifyErrorIfExists(errors.New("failed to load relayer channel. Please make sure that the rollapp is " +
-					"running on your local machine and a channel has been established"))
+					"running on your local machine and a relayer channel has been established"))
 			}
-			fundFuacetCmd := exec.Command(rlpCfg.RollappBinary, "tx", "ibc-transfer", "transfer", "transfer",
+			fundFaucetCmd := exec.Command(rlpCfg.RollappBinary, "tx", "ibc-transfer", "transfer", "transfer",
 				srcChannel, "dym1g8sf7w4cz5gtupa6y62h3q6a4gjv37pgefnpt5", "5000000000000000000000000"+rlpCfg.Denom, "--from",
 				consts.KeysIds.RollappSequencer, "--keyring-backend", "test", "--home", filepath.Join(rlpCfg.Home,
 					consts.ConfigDirName.Rollapp), "--broadcast-mode", "block")
-			stdout, err := utils.ExecBashCommandWithStdout(fundFuacetCmd)
+			stdout, err := utils.ExecBashCommandWithStdout(fundFaucetCmd)
+			utils.PrettifyErrorIfExists(err)
+			err = tx_utils.CheckTxStdOut(stdout)
 			utils.PrettifyErrorIfExists(err)
 			fmt.Println("💈 The Dymension faucet has been funded successfully!")
 		},
