+++ cmd/tx/register/bash_commands.go
